+++ backend/src/routes/categories.js
@@ -5,9 +5,9 @@ import {
   updateCategory,
   deleteCategory,
 } from '../controllers/categoryController.js';
-import { categoryValidator } from '../validators/categoryValidators.js';
 import { validate } from '../middleware/validate.js';
-import { IdValidator } from '../validators/IdValidator.js';
+import { categoryValidator } from '../validators/validateCategory.js';
+import { IdValidator } from '../validators/validateId.js';
 
 const router = express.Router();
