+++ backend/src/validators/validateCategory.js
